requirements.txt
# requirements for the model_training.py file
argparse==1.4.0
pandas==1.4.2
numpy==1.22.4
joblib==1.1.1
# os-sys==2.1.4
scikit-learn==0.24.2    # auto-sklearn 0.15.0 requires scikit version >= 0.24.0 and < 0.25.0
auto-sklearn==0.14.7


# requirements for the prediction.py file

influxdb==5.3.1
iso8601==1.1.0
python-daemon==3.0.1


